+++ familyhelper-clannad-impl/src/test/java/com/dwarfeng/familyhelper/clannad/impl/configuration/ServiceExceptionMapperConfiguration.java
@@ -2,6 +2,7 @@ package com.dwarfeng.familyhelper.clannad.impl.configuration;
 
 import com.dwarfeng.familyhelper.clannad.sdk.util.ServiceExceptionCodes;
 import com.dwarfeng.familyhelper.clannad.stack.exception.AvatarNotExistsException;
+import com.dwarfeng.familyhelper.clannad.stack.exception.NotificationNotExistsException;
 import com.dwarfeng.familyhelper.clannad.stack.exception.ProfileNotExistsException;
 import com.dwarfeng.familyhelper.clannad.stack.exception.UserNotExistsException;
 import com.dwarfeng.subgrade.impl.exception.MapServiceExceptionMapper;
@@ -22,6 +23,7 @@ public class ServiceExceptionMapperConfiguration {
         destination.put(ProfileNotExistsException.class, ServiceExceptionCodes.PROFILE_NOT_EXISTS);
         destination.put(UserNotExistsException.class, ServiceExceptionCodes.USER_NOT_EXISTS);
         destination.put(AvatarNotExistsException.class, ServiceExceptionCodes.AVATAR_NOT_EXISTS);
+        destination.put(NotificationNotExistsException.class, ServiceExceptionCodes.NOTIFICATION_NOT_EXISTS);
         return new MapServiceExceptionMapper(destination, com.dwarfeng.subgrade.sdk.exception.ServiceExceptionCodes.UNDEFINE);
     }
 }