+++ familyhelper-clannad-sdk/src/main/java/com/dwarfeng/familyhelper/clannad/sdk/util/ServiceExceptionCodes.java
@@ -18,6 +18,8 @@ public final class ServiceExceptionCodes {
             new ServiceException.Code(offset(10), "user not exists");
     public static final ServiceException.Code AVATAR_NOT_EXISTS =
             new ServiceException.Code(offset(20), "avatar not exists");
+    public static final ServiceException.Code NOTIFICATION_NOT_EXISTS =
+            new ServiceException.Code(offset(30), "notification not exists");
 
     private static int offset(int i) {
         return EXCEPTION_CODE_OFFSET + i;
@@ -45,6 +47,7 @@ public final class ServiceExceptionCodes {
         PROFILE_NOT_EXISTS.setCode(offset(0));
         USER_NOT_EXISTS.setCode(offset(10));
         AVATAR_NOT_EXISTS.setCode(offset(20));
+        NOTIFICATION_NOT_EXISTS.setCode(offset(30));
     }
 
     private ServiceExceptionCodes() {